fix(cloud): satisfy managed sandbox CI checks
