+++ server/proliferate/integrations/github/__init__.py
@@ -1,21 +1,6 @@
 """Public GitHub integration API."""
 
 from proliferate.integrations.github import repos as _repos
-from proliferate.integrations.github.repos import (
-    GitHubIntegrationError,
-    GitHubInvalidCursor,
-    GitHubRateLimited,
-    GitHubRepoAccessRequired,
-    GitHubRepoBranches,
-    GitHubRepoEmpty,
-    GitHubRepositoryPage,
-    GitHubRepositorySummary,
-    GitHubUserProfile,
-    get_github_repo_branches,
-    get_github_user_profile,
-    list_branches,
-    list_github_repositories,
-)
 from proliferate.integrations.github.app_installations import (
     GitHubAppInstallationInfo,
     GitHubAppRepositoryCoverage,
@@ -30,6 +15,21 @@ from proliferate.integrations.github.app_user_tokens import (
     GitHubAppUserAuthorization,
     exchange_github_app_code,
     refresh_github_app_user_authorization,
+)
+from proliferate.integrations.github.repos import (
+    GitHubIntegrationError,
+    GitHubInvalidCursor,
+    GitHubRateLimited,
+    GitHubRepoAccessRequired,
+    GitHubRepoBranches,
+    GitHubRepoEmpty,
+    GitHubRepositoryPage,
+    GitHubRepositorySummary,
+    GitHubUserProfile,
+    get_github_repo_branches,
+    get_github_user_profile,
+    list_branches,
+    list_github_repositories,
 )
 
 httpx = _repos.httpx
